+++ CHANGES.md
@@ -58,6 +58,7 @@ Changes in 12.0
 - Close #348 Add School Year & Principal to Substitutions in Letters.php
 - Give full responsibility to & improve AllowEdit() & AllowUse() in AllowEdit.fnc.php
 - Add smart cache: up to 8ms faster in AllowEdit.fnc.php
+- Rely solely on AllowUse() in Menu.php, Modules.php & Bottom.php
 
 Changes in 11.8.4
 -----------------